requirements.txt
psycopg2-binary
python-dotenv
openai
qdrant-client
sqlglot
pandas
tabulate
pypdf
langchain
openai-agents
asyncio
redis
fastapi
uvicorn
fastmcp[cli]
fastmcp